feat: manage users(crud) Integrations UI
